src/backend/requirements.txt
# Core dependencies
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-multipart>=0.0.6,<1.0.0

# Azure SDK
azure-identity>=1.15.0,<2.0.0
azure-storage-blob>=12.19.0,<13.0.0
azure-cosmos>=4.5.0,<5.0.0
azure-search-documents>=11.4.0,<12.0.0
openai>=1.10.0,<2.0.0

# Observability
opentelemetry-instrumentation-fastapi>=0.43b0
azure-monitor-opentelemetry>=1.2.0,<2.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
httpx>=0.26.0,<1.0.0
